+++ pcs_test/tier0/lib/corosync/test_config_validators_create.py
@@ -1204,7 +1204,14 @@ class CreateLinkListKnet(CreateLinkListCommonMixin, TestCase):
                 ],
                 2,
             ),
-            [],
+            [
+                fixture.deprecation(
+                    report_codes.DEPRECATED_OPTION_VALUE,
+                    option_name="transport",
+                    deprecated_value="sctp",
+                    replaced_by=None,
+                )
+            ],
         )
 
     def test_invalid_all_values(self):