+++ pcs_test/tier0/lib/corosync/test_config_validators_links.py
@@ -765,6 +765,29 @@ class AddLink(TestCase):
             ],
         )
 
+    def test_deprecated_sctp_knet_transport(self):
+        assert_report_item_list_equal(
+            config_validators.add_link(
+                self.new_addrs,
+                {
+                    "transport": "sctp",
+                },
+                self.coro_nodes,
+                self.pcmk_nodes,
+                self.existing_link_list,
+                self.transport,
+                constants.IP_VERSION_64,
+            ),
+            [
+                fixture.deprecation(
+                    report_codes.DEPRECATED_OPTION_VALUE,
+                    option_name="transport",
+                    deprecated_value="sctp",
+                    replaced_by=None,
+                )
+            ],
+        )
+
 
 class RemoveLinks(TestCase):
     def setUp(self):
@@ -1833,6 +1856,33 @@ class UpdateLinkKnet(TestCase):
                     allowed_values=("sctp", "udp"),
                     cannot_be_empty=False,
                     forbidden_characters=None,
+                ),
+            ],
+        )
+
+    def test_deprecated_sctp_knet_transport(self):
+        assert_report_item_list_equal(
+            config_validators.update_link(
+                "2",
+                {},
+                {
+                    "transport": "sctp",
+                },
+                {},
+                [],
+                [],
+                [
+                    "2",
+                ],
+                constants.TRANSPORTS_KNET[0],
+                constants.IP_VERSION_64,
+            ),
+            [
+                fixture.deprecation(
+                    report_codes.DEPRECATED_OPTION_VALUE,
+                    option_name="transport",
+                    deprecated_value="sctp",
+                    replaced_by=None,
                 ),
             ],
         )