+++ CHANGELOG.md
@@ -6,7 +6,12 @@
 - Do not wrap resource and stonith agent description to preserve existing
   formatting ([RHEL-113763])
 
+### Deprecated
+- Value `sctp` of the knet link option `transport` is deprecated in
+  corosync / knet and might be removed in a future release ([RHEL-126842])
+
 [RHEL-113763]: https://issues.redhat.com/browse/RHEL-113763
+[RHEL-126842]: https://issues.redhat.com/browse/RHEL-126842
 
 
 ## [0.11.10] - 2025-07-09